Feature/contract

Approved-by: Rachael Hong <rachael@r2v.io>

+++ yggdrash-core/src/test/java/io/yggdrash/core/contract/RuntimeTest.java
@@ -31,16 +31,17 @@ import org.slf4j.LoggerFactory;
 import org.spongycastle.crypto.InvalidCipherTextException;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class RuntimeTest {
     private static final Logger log = LoggerFactory.getLogger(RuntimeTest.class);
     private final TransactionReceiptStore txReceiptStore = new TransactionReceiptStore();
-    private final Contract<Long> coinContract = new CoinContract();
+    private final Contract<BigDecimal> coinContract = new CoinContract();
     private final Contract<JsonObject> stemContract = new StemContract();
     private io.yggdrash.core.contract.Runtime<JsonObject> stemRuntime;
-    private io.yggdrash.core.contract.Runtime<Long> yeedRuntime;
+    private io.yggdrash.core.contract.Runtime<BigDecimal> yeedRuntime;
     private Wallet wallet;
     private BranchId branchId;
